Day 3 oop02_03

+++ src/ch/teko/prg2/tag02/Work/ue_oop_02_02/KonzertTicket.java
@@ -1,9 +1,9 @@
 package ch.teko.prg2.tag02.Work.ue_oop_02_02;
 
 public class KonzertTicket extends Ticket{
-    private double sitzreihe;
+    private int sitzreihe;
 
-    public KonzertTicket(Ort eventOrt, String eventName, int basisPreis, double sitzreihe){
+    public KonzertTicket(Ort eventOrt, String eventName, int basisPreis, int sitzreihe){
         super(eventOrt, eventName, basisPreis);
         this.sitzreihe=sitzreihe;
     }
@@ -15,7 +15,7 @@ public class KonzertTicket extends Ticket{
     @Override
     public double berechneTicketPreis() {
         double berechneticketpreis;
-        berechneticketpreis =this.getBasisPreis() * (1 + 1/this.getsitzreihe());
+        berechneticketpreis =this.getBasisPreis() * (1 + 1.0/this.getsitzreihe());
         this.setTicketPreis(berechneticketpreis);
         return this.getTicketPreis();
     }
